src/main/java/org/kakaoshare/backend/jwt/service/UserDetailsServiceImpl.java
package org.kakaoshare.backend.jwt.service;

import lombok.RequiredArgsConstructor;
import org.kakaoshare.backend.domain.member.entity.MemberDetails;
import org.kakaoshare.backend.domain.member.repository.MemberRepository;
import org.springframework.security.core.userdetails.UserDetails;
import org.springframework.security.core.userdetails.UserDetailsService;
import org.springframework.security.core.userdetails.UsernameNotFoundException;
import org.springframework.stereotype.Service;
import org.springframework.transaction.annotation.Transactional;

@RequiredArgsConstructor
@Service
@Transactional(readOnly = true)
public class UserDetailsServiceImpl implements UserDetailsService {
    private final MemberRepository memberRepository;

    @Override
    public UserDetails loadUserByUsername(final String username) throws UsernameNotFoundException {
        return memberRepository.findByProviderId(username)
                .map(MemberDetails::new)
                .orElseThrow(() -> new IllegalArgumentException("유효하지 않은 이메일 입니다."));
    }
}
